requirements.txt
google-analytics-data
google-analytics-admin
python-dateutil
Jinja2
